backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
yfinance==0.2.40
pydantic==2.9.0
python-dateutil==2.9.0
openpyxl==3.1.5
python-dotenv==1.2.1
requests>=2.28.0
pyotp>=2.9.0
mcp>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
PyJWT>=2.8.0
google-auth>=2.29.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0
